requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Document Processing
python-docx==1.1.0

# AI/LLM Integration
openai==1.3.8

# Evaluation Metrics
nltk==3.8.1
rouge-score==0.1.2
sacrebleu==2.3.1
comet-ml==3.35.5
unbabel-comet==2.2.2

# Text Processing
spacy==3.7.2
scikit-learn==1.3.2
numpy==1.24.3
scipy==1.11.4

# Experiment Tracking
mlflow==2.8.1

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
pydantic-settings==2.1.0

# Development/Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Security
passlib==1.7.4
python-jose[cryptography]==3.3.0
pypdf2==3.0.1
